# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JJK Mission Mailer API",
        "description": "API for managing sorcerer missions and assignments",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/mission/{id}": {
            "get": {
                "tags": [
                    "Missions"
                ],
                "summary": "Get a mission by ID",
                "description": "Returns a single mission with curses and assignments",
                "operationId": "ebdeaea0a06e07aa77fefa53cd411c59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mission details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Mission"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mission not found"
                    }
                }
            }
        },
        "/mission/generate": {
            "post": {
                "tags": [
                    "Missions"
                ],
                "summary": "Generate a new mission",
                "description": "Creates a new mission based on provided parameters. Optionally links a curse and assigns a sorcerer.",
                "operationId": "8241c121ffbce6b938ba0ad1d0b57ed0",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sorcerer_id": {
                                        "description": "ID of sorcerer to assign",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "curse_id": {
                                        "description": "ID of curse to link",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "curse_level": {
                                        "type": "string",
                                        "enum": [
                                            "grade-4",
                                            "grade-3",
                                            "grade-2",
                                            "semi-grade-1",
                                            "grade-1",
                                            "special-grade"
                                        ]
                                    },
                                    "sorcerer_grade": {
                                        "type": "string",
                                        "enum": [
                                            "grade-4",
                                            "grade-3",
                                            "grade-2",
                                            "semi-grade-1",
                                            "grade-1",
                                            "special-grade"
                                        ]
                                    },
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "location": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "urgency_level": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mission created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mission generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "mission": {
                                                    "$ref": "#/components/schemas/Mission"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/mission/{id}/send-email": {
            "post": {
                "tags": [
                    "Missions"
                ],
                "summary": "Send mission emails to assigned sorcerers",
                "description": "Sends email notifications to all sorcerers assigned to this mission",
                "operationId": "b9a47980c2369ac7d9952c1041a63c5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Emails sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "2 email(s) sent"
                                        },
                                        "mission": {
                                            "$ref": "#/components/schemas/Mission"
                                        },
                                        "notified_sorcerers": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Sorcerer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No sorcerers assigned to this mission"
                    },
                    "404": {
                        "description": "Mission not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Mission": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Exorcise: Finger Bearer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "required_sorcerer_grade": {
                        "type": "string",
                        "example": "grade-2"
                    },
                    "curse_level": {
                        "type": "string",
                        "example": "grade-2"
                    },
                    "category": {
                        "type": "string",
                        "example": "exorcism"
                    },
                    "location": {
                        "type": "string",
                        "example": "Tokyo"
                    },
                    "urgency_level": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "available"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Sorcerer": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Yuji Itadori"
                    },
                    "email": {
                        "type": "string",
                        "example": "yuji@jujutsu.jp"
                    },
                    "sorcerer_grade": {
                        "type": "string",
                        "example": "grade-1"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            }
        }
    },
...
}